pyproject.toml
[build-system]
requires = ["hatchling>=1.4.1"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "dotbot/frontend/*",
  "dotbot/lib/*",
  "*.py"
]

[tool.hatch.build.targets.sdist.hooks.custom]
path = "utils/hooks/build.py"

[project]
name = "pydotbot"
version = "0.7.0"
authors = [
    { name="Alexandre Abadie", email="alexandre.abadie@inria.fr" },
    { name="Theo Akbas", email="theo.akbas@inria.fr" },
]
dependencies = [
    "click          == 8.1.3",
    "fastapi        == 0.85.1",
    "numpy          == 1.23.4",
    "opencv-python  == 4.6.0.66",
    "pygame         == 2.1.2",
    "pyserial       == 3.5",
    "pynput         == 1.7.6",
    "rich           == 12.6.0",
    "uvicorn        == 0.18.3",
    "websockets     == 10.4",
]
description = "Package to easily control your DotBots and SailBots."
readme = "README.md"
license = { text="MIT" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/DotBots/BotController-python"
"Bug Tracker" = "https://github.com/DotBots/BotController-python/issues"

[project.scripts]
dotbot-controller = "dotbot.main:main"
